fix bad check on batch size in tests
